Add confirmation messages

+++ clean.sh
@@ -2,5 +2,8 @@
 
 . ./init-env.sh
 
-docker-compose -f docker-compose.yml -f docker-compose.dev.yml down -v --rmi 'all' --remove-orphans
-docker-compose -f docker-compose.yml -f docker-compose.dev.yml rm -v
+read -p "Are you sure to delete containers AND volumes AND images [y/n]? " -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]; then
+  docker-compose -f docker-compose.yml -f docker-compose.dev.yml down -v --rmi 'all' --remove-orphans
+  docker-compose -f docker-compose.yml -f docker-compose.dev.yml rm -v
+fi
